+++ components/PageWrapper/index.tsx
@@ -1,16 +1,22 @@
 import React, { useEffect, ReactNode } from 'react';
-
+import BackButton from "@/components/BackButton";
 interface PageWrapperProps {
     title: string;
+    backButton: boolean;
     children: ReactNode;
 }
 
-const PageWrapper: React.FC<PageWrapperProps> = ({ title, children }) => {
+const PageWrapper: React.FC<PageWrapperProps> = ({ title, backButton, children }) => {
     useEffect(() => {
         document.title = `${title} | DMPTool`;
         window.scrollTo(0, 0);
     }, [title]);
-    return <>{children}</>;
+    return (
+        <>
+            <BackButton />
+            {children}
+        </>
+    )
 }
 
 export default PageWrapper;